Add Support for -MD, -MMD, -gccdep
Makes `Compiler` aware of any dependency file that `mwccpsp.exe` might
emit and if run through `wibo`, it converts the paths from DOS to Unix
paths. This allows tools like `make` or `ninja` to rebuild objects based
on the source file's embedded dependencies ensuring that files get
rebuilt when necessary and avoiding rebuilding files when not needed.

`mwccpsp.exe` supports two modes when emitting dependency files in
addition to a compiled output - gcc compatibility mode (specified by the
`-gccdep` argument) where the file is placed next to the output object,
and another where the file is placed in `$PWD`. Both modes are
supported, but MetroWorks mode only works with a regular file input (and
not stdin).

In cases where the input C file is read from stdin, the source
dependency for the make rule cannot be determined. Instead of using the
ephemeral C file, the dependency is removed from the output rule.

In all cases the include headers that are emitted are absoluate paths,
unlike GCC. Additional post-processing could be done to determine if the
absolute path matches one of the relative include paths, but that is not
necessary for the current use case.

`-M` and `-MM` options are not supported because they change the primary
output of `mwccpsp.exe` and later logic assumes its processing an ELF
file. This case could be detected as well, but is not currently needed.